loadbalancers: Support nodes where nodeName is the private or public IP
Fix #69
